@@ -57,91 +59,91 @@
   },
   "devDependencies": {
     "@angular-devkit/build-optimizer": "0.0.32",
-    "@angular/tsc-wrapped": "^4.4.6",
+    "@angular/tsc-wrapped": "4.4.6",
     "@ngtools/webpack": "1.8.0",
     "@types/angular": "1.6.6",
-    "@types/angular-ui-router": "^1.1.36",
-    "@types/chance": "^0.7.33",
-    "@types/chart.js": "^2.4.6",
-    "@types/jasmine": "^2.5.53",
-    "@types/jquery": "^2.0.41",
-    "@types/moment-timezone": "^0.2.34",
-    "@types/node": "^7.0.15",
-    "@types/prismjs": "^1.6.2",
-    "@types/webpack-env": "^1.13.0",
-    "angular-mocks": "~1.3.2",
-    "angular-router-loader": "^0.6.0",
-    "angular2-template-loader": "^0.6.2",
-    "awesome-typescript-loader": "^3.3.0",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.1",
+    "@types/angular-ui-router": "1.1.40",
+    "@types/chance": "0.7.35",
+    "@types/chart.js": "2.6.11",
+    "@types/dragula": "2.1.33",
+    "@types/jasmine": "2.6.2",
+    "@types/jquery": "2.0.48",
+    "@types/moment-timezone": "0.2.35",
+    "@types/node": "7.0.46",
+    "@types/prismjs": "1.6.5",
+    "@types/webpack-env": "1.13.2",
+    "angular-mocks": "1.3.2",
+    "angular-router-loader": "0.6.0",
+    "angular2-template-loader": "0.6.2",
+    "awesome-typescript-loader": "3.3.0",
+    "babel-core": "6.26.0",
+    "babel-loader": "7.1.2",
+    "babel-preset-env": "1.6.1",
     "canonical-path": "0.0.2",
-    "chromedriver": "^2.36.0",
-    "concurrently": "^3.5.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.7",
-    "expose-loader": "^0.7.3",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.5",
-    "graceful-fs": "^4.1.11",
-    "grunt": "^1.0.1",
-    "grunt-banner": "^0.6.0",
-    "grunt-contrib-clean": "^1.1.0",
-    "grunt-contrib-compress": "^1.4.3",
-    "grunt-contrib-concat": "^1.0.1",
-    "grunt-contrib-connect": "^1.0.2",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-cssmin": "^2.2.0",
-    "grunt-contrib-jasmine": "^1.1.0",
-    "grunt-contrib-jshint": "^1.1.0",
-    "grunt-contrib-less": "^1.4.1",
-    "grunt-contrib-rename": "^0.2.0",
-    "grunt-contrib-uglify": "^2.3.0",
-    "grunt-execute": "^0.2.2",
-    "grunt-istanbul": "~0.6.1",
-    "grunt-protractor-runner": "^5.0.0",
-    "grunt-run": "^0.7.0",
-    "grunt-stylelint": "^0.8.0",
-    "grunt-ts": "^6.0.0-beta.16",
-    "grunt-tslint": "^5.0.1",
-    "grunt-webfont": "^0.4.8",
-    "grunt-webpack": "^3.0.0",
-    "html-loader": "^0.5.1",
-    "html-webpack-plugin": "^2.30.1",
-    "jasmine": "^2.6.0",
-    "jasmine-reporters": "^2.2.1",
-    "jit-grunt": "^0.10.0",
-    "jshint": "^2.9.5",
-    "jshint-loader": "^0.8.4",
-    "json-loader": "^0.5.7",
-    "less-loader": "^4.0.5",
-    "lite-server": "^2.3.0",
-    "load-grunt-config": "^0.19.2",
-    "lodash": "^4.17.4",
-    "markdown-code-highlight-loader": "^1.0.1",
-    "marked": "^0.3.6",
+    "chromedriver": "2.36.0",
+    "concurrently": "3.5.1",
+    "copy-webpack-plugin": "4.2.0",
+    "css-loader": "0.28.7",
+    "expose-loader": "0.7.4",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.5",
+    "graceful-fs": "4.1.11",
+    "grunt": "1.0.1",
+    "grunt-banner": "0.6.0",
+    "grunt-contrib-clean": "1.1.0",
+    "grunt-contrib-compress": "1.4.3",
+    "grunt-contrib-concat": "1.0.1",
+    "grunt-contrib-connect": "1.0.2",
+    "grunt-contrib-copy": "1.0.0",
+    "grunt-contrib-cssmin": "2.2.1",
+    "grunt-contrib-jasmine": "1.1.0",
+    "grunt-contrib-jshint": "1.1.0",
+    "grunt-contrib-less": "1.4.1",
+    "grunt-contrib-rename": "0.2.0",
+    "grunt-contrib-uglify": "2.3.0",
+    "grunt-execute": "0.2.2",
+    "grunt-istanbul": "0.6.1",
+    "grunt-protractor-runner": "5.0.0",
+    "grunt-run": "0.7.0",
+    "grunt-stylelint": "0.8.0",
+    "grunt-ts": "6.0.0-beta.17",
+    "grunt-tslint": "5.0.1",
+    "grunt-webfont": "0.4.8",
+    "grunt-webpack": "3.0.2",
+    "html-loader": "0.5.1",
+    "html-webpack-plugin": "2.30.1",
+    "jasmine": "2.8.0",
+    "jasmine-reporters": "2.2.1",
+    "jit-grunt": "0.10.0",
+    "jshint": "2.9.5",
+    "jshint-loader": "0.8.4",
+    "less-loader": "4.0.5",
+    "lite-server": "2.3.0",
+    "load-grunt-config": "0.19.2",
+    "lodash": "4.17.4",
+    "markdown-code-highlight-loader": "1.0.1",
+    "marked": "0.3.6",
     "ng-cache-loader": "0.0.26",
-    "ng-packagr": "^1.5.1",
-    "null-loader": "^0.1.1",
-    "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "ng-packagr": "1.6.0",
+    "null-loader": "0.1.1",
+    "optimize-css-assets-webpack-plugin": "3.2.0",
     "prismjs-loader": "0.0.4",
-    "protractor": "^5.3.0",
-    "protractor-istanbul-plugin": "^2.0.0",
+    "protractor": "5.3.0",
+    "protractor-istanbul-plugin": "2.0.0",
     "protractor-jasmine2-html-reporter": "0.0.7",
-    "raw-loader": "^0.5.1",
-    "script-loader": "^0.7.0",
-    "style-loader": "^0.18.2",
-    "stylelint": "^7.12.0",
-    "time-grunt": "^1.4.0",
-    "to-string-loader": "^1.1.5",
-    "tslint": "^5.4.3",
+    "raw-loader": "0.5.1",
+    "script-loader": "0.7.2",
+    "style-loader": "0.18.2",
+    "stylelint": "7.13.0",
+    "time-grunt": "1.4.0",
+    "to-string-loader": "1.1.5",
+    "tslint": "5.8.0",
     "typescript": "2.4.2",
-    "uglify-loader": "^2.0.0",
-    "uglifyjs-webpack-plugin": "^1.0.1",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.5.0"
+    "uglify-js": "3.3.15",
+    "uglifyjs-webpack-plugin": "1.1.6",
+    "url-loader": "0.6.2",
+    "webpack": "3.10.0",
+    "webpack-dev-server": "2.9.3"
   },
   "files": [
     "configs/loaders",
