@@ -1,15 +1,20 @@
-import { NgModule, ComponentFactoryResolver } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ComponentFactoryResolver, NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { DocumentationComponentsModule } from '../../../../components/components.module';
-import { ResolverService, DocumentationPage } from '../../../../services/resolver/resolver.service';
-import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
-
-import { ComponentsComponentListNg1Component } from './component-list-ng1/component-list-ng1.component';
-import { WrappersModule } from '../../../../wrappers/wrappers.module';
 import { TabsModule } from 'ngx-bootstrap/tabs';
 
+import { DocumentationComponentsModule } from '../../../../components/components.module';
+import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
+import { DocumentationPage, ResolverService } from '../../../../services/resolver/resolver.service';
+import { WrappersModule } from '../../../../wrappers/wrappers.module';
+import { ComponentsComponentListNg1Component } from './component-list-ng1/component-list-ng1.component';
+import { ComponentsComponentListComponent } from './component-list/component-list.component';
+import { FocusIfModule } from '../../../../../../src';
+
 const SECTIONS = [
-    ComponentsComponentListNg1Component
+    ComponentsComponentListNg1Component,
+    ComponentsComponentListComponent
 ];
 
 const ROUTES = [
@@ -24,8 +29,12 @@ const ROUTES = [
 
 @NgModule({
     imports: [
+        CommonModule,
         WrappersModule,
         TabsModule,
+        FormsModule,
+        FocusIfModule,
+        ReactiveFormsModule,
         DocumentationComponentsModule,
         RouterModule.forChild(ROUTES)
     ],
