@@ -1,14 +1,18 @@
 import { NgModule, ComponentFactoryResolver } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { AccordionModule } from 'ngx-bootstrap/accordion';
 import { ModalModule } from 'ngx-bootstrap/modal';
+import { TabsModule } from 'ngx-bootstrap/tabs';
+import { EboxModule, SidePanelModule, CheckboxModule, RadioButtonModule } from '../../../../../../src/index';
+import { ItemDisplayPanelModule } from '../../../../../../src/components/item-display-panel/index';
 
-import { ComponentsItemDisplayPanelInlineComponent } from './item-display-panel-inline/item-display-panel-inline.component';
 import { SparkModule } from './../../../../../../src/components/spark/spark.module';
 import { DocumentationComponentsModule } from '../../../../components/components.module';
 import { ResolverService, DocumentationPage } from '../../../../services/resolver/resolver.service';
 import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
+import { WrappersModule } from '../../../../wrappers/wrappers.module';
 
 import { ComponentsAccordionNg1Component } from './accordion-ng1/accordion-ng1.component';
 import { ComponentsItemDisplayPanelNg1Component } from './item-display-panel-ng1/item-display-panel-ng1.component';
@@ -16,12 +20,10 @@ import { ComponentsItemDisplayPanelServiceNg1Component } from './item-display-pa
 import { ComponentsModalInsetPanelNg1Component } from './modal-inset-panel-ng1/item-display-panel-ng1.component';
 import { ComponentsSideInsetPanelNg1Component } from './side-inset-panel-ng1/side-inset-panel-ng1.component';
 import { ComponentsItemDisplayPanelComponent } from './item-display-panel/item-display-panel.component';
-import { WrappersModule } from '../../../../wrappers/wrappers.module';
-import { TabsModule } from 'ngx-bootstrap/tabs';
+import { ComponentsItemDisplayPanelInlineComponent } from './item-display-panel-inline/item-display-panel-inline.component';
 import { ComponentsEboxComponent } from './ebox/ebox.component';
-import { EboxModule } from '../../../../../../src/index';
-import { ItemDisplayPanelModule } from '../../../../../../src/components/item-display-panel/index';
 import { ComponentsAccordionComponent } from './accordion/accordion.component';
+import { ComponentsSidePanelComponent } from './side-panel/side-panel.component';
 
 const SECTIONS = [
     ComponentsAccordionComponent,
@@ -32,6 +34,7 @@ const SECTIONS = [
     ComponentsItemDisplayPanelInlineComponent,
     ComponentsModalInsetPanelNg1Component,
     ComponentsSideInsetPanelNg1Component,
+    ComponentsSidePanelComponent,
     ComponentsEboxComponent
 ];
 
@@ -47,16 +50,20 @@ const ROUTES = [
 
 @NgModule({
     imports: [
-        WrappersModule,
-        TabsModule,
-        EboxModule,
-        DocumentationComponentsModule,
         CommonModule,
-        ItemDisplayPanelModule,
-        ModalModule.forRoot(),
+        FormsModule,
         RouterModule.forChild(ROUTES),
+        AccordionModule.forRoot(),
+        ModalModule.forRoot(),
+        DocumentationComponentsModule,
+        CheckboxModule,
+        EboxModule,
+        ItemDisplayPanelModule,
+        RadioButtonModule,
+        SidePanelModule,
         SparkModule,
-        AccordionModule.forRoot()
+        TabsModule,
+        WrappersModule,
     ],
     exports: SECTIONS,
     declarations: SECTIONS,
