@@ -1,10 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
 import { BaseDocumentationSection } from '../../../../../components/base-documentation-section/base-documentation-section';
 
 @Component({
     selector: 'uxd-components-dynamic-filters',
-    templateUrl: './dynamic-filters-ng1.component.html'
+    templateUrl: './dynamic-filters-ng1.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 @DocumentationSectionComponent('ComponentsDynamicFiltersNg1Component')
 export class ComponentsDynamicFiltersNg1Component extends BaseDocumentationSection {
