@@ -1,12 +1,13 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { BaseDocumentationSection } from '../../../../../components/base-documentation-section/base-documentation-section';
 
 @Component({
-    selector: 'uxd-components-reorderable-table',
-    templateUrl: './reorderable-table-ng1.component.html'
+    selector: 'uxd-components-reorderable-table-ng1',
+    templateUrl: './reorderable-table-ng1.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 @DocumentationSectionComponent('ComponentsReorderableTableNg1Component')
 export class ComponentsReorderableTableNg1Component extends BaseDocumentationSection implements ICodePenProvider {
