@@ -1,10 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
 import { BaseDocumentationSection } from '../../../../../components/base-documentation-section/base-documentation-section';
 
 @Component({
-    selector: 'uxd-components-sorting',
-    templateUrl: './sorting-ng1.component.html'
+    selector: 'uxd-components-sorting-ng1',
+    templateUrl: './sorting-ng1.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 @DocumentationSectionComponent('ComponentsSortingNg1Component')
 export class ComponentsSortingNg1Component extends BaseDocumentationSection {
