@@ -1,21 +1,26 @@
-import { NgModule, ComponentFactoryResolver } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ComponentFactoryResolver, NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { AccordionModule } from 'ngx-bootstrap/accordion';
+import { TabsModule } from 'ngx-bootstrap/tabs';
+import { CardTabsModule, RadioButtonModule } from '../../../../../../src';
 import { DocumentationComponentsModule } from '../../../../components/components.module';
-import { ResolverService, DocumentationPage } from '../../../../services/resolver/resolver.service';
 import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
-
-import { ComponentsTabsNg1Component } from './tabs-ng1/tabs-ng1.component';
+import { DocumentationPage, ResolverService } from '../../../../services/resolver/resolver.service';
+import { WrappersModule } from '../../../../wrappers/wrappers.module';
+import { ComponentsCardTabsNg1Component } from './card-tabs-ng1/card-tabs-ng1.component';
+import { ComponentsCardTabsComponent } from './card-tabs/card-tabs.component';
 import { ComponentsDetailedTabExampleNg1Component } from './detailed-tab-example-ng1/detailed-tab-example-ng1.component';
 import { ComponentsStackedTabsNg1Component } from './stacked-tabs-ng1/stacked-tabs-ng1-component';
-import { ComponentsCardTabsNg1Component } from './card-tabs-ng1/card-tabs-ng1.component';
-import { WrappersModule } from '../../../../wrappers/wrappers.module';
-import { TabsModule } from 'ngx-bootstrap/tabs';
+import { ComponentsTabsNg1Component } from './tabs-ng1/tabs-ng1.component';
+
 
 const SECTIONS = [
     ComponentsTabsNg1Component,
     ComponentsDetailedTabExampleNg1Component,
     ComponentsStackedTabsNg1Component,
-    ComponentsCardTabsNg1Component
+    ComponentsCardTabsNg1Component,
+    ComponentsCardTabsComponent
 ];
 
 const ROUTES = [
@@ -30,8 +35,12 @@ const ROUTES = [
 
 @NgModule({
     imports: [
+        CommonModule,
         WrappersModule,
         TabsModule,
+        CardTabsModule,
+        RadioButtonModule,
+        AccordionModule.forRoot(),
         DocumentationComponentsModule,
         RouterModule.forChild(ROUTES)
     ],
