@@ -1,21 +1,25 @@
-import { NgModule, ComponentFactoryResolver } from '@angular/core';
+import { ComponentFactoryResolver, NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { DocumentationComponentsModule } from '../../../../components/components.module';
-import { ResolverService, DocumentationPage } from '../../../../services/resolver/resolver.service';
-import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
+import { TabsModule } from 'ngx-bootstrap/tabs';
+import { PopoverModule } from 'ngx-bootstrap/popover';
 
-import { ComponentsTooltipsNg1Component } from './tooltips-ng1/tooltips-ng1.component';
+import { DocumentationComponentsModule } from '../../../../components/components.module';
+import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
+import { DocumentationPage, ResolverService } from '../../../../services/resolver/resolver.service';
+import { WrappersModule } from '../../../../wrappers/wrappers.module';
 import { ComponentsOverflowTooltipNg1Component } from './overflow-tooltip-ng1/overflow-tooltip-ng1.component';
 import { ComponentsSingleLineOverflowTooltipNg1Component } from './single-line-overflow-tooltip-ng1/single-line-overflow-tooltip-ng1.component';
 import { ComponentsStaticTooltipNg1Component } from './static-tooltip-ng1/static-tooltip-ng1.component';
-import { WrappersModule } from '../../../../wrappers/wrappers.module';
-import { TabsModule } from 'ngx-bootstrap/tabs';
+import { ComponentsStaticTooltipComponent } from './static-tooltip/static-tooltip.component';
+import { ComponentsTooltipsNg1Component } from './tooltips-ng1/tooltips-ng1.component';
+
 
 const SECTIONS = [
     ComponentsTooltipsNg1Component,
     ComponentsOverflowTooltipNg1Component,
     ComponentsSingleLineOverflowTooltipNg1Component,
-    ComponentsStaticTooltipNg1Component
+    ComponentsStaticTooltipNg1Component,
+    ComponentsStaticTooltipComponent
 ];
 
 const ROUTES = [
@@ -32,6 +36,7 @@ const ROUTES = [
     imports: [
         WrappersModule,
         TabsModule,
+        PopoverModule.forRoot(),
         DocumentationComponentsModule,
         RouterModule.forChild(ROUTES)
     ],
