@@ -1,20 +1,15 @@
-import { browser } from 'protractor';
+import { browser, ElementFinder, ElementArrayFinder } from 'protractor';
 import { PageHeaderPage } from './page-header.po.spec';
 
 describe('Page Header Tests', () => {
 
   let page: PageHeaderPage;
-  let browserName: string;
 
   beforeEach(() => {
     page = new PageHeaderPage();
     page.getPage();
     
     page.condensed = false;
-    
-    browser.getCapabilities().then(function(caps) {
-        browserName = caps.get('browserName');
-    });
   });
 
   it('should have correct initial states', () => {
@@ -119,5 +114,142 @@ describe('Page Header Tests', () => {
     page.openActions();
     expect<any>(page.confirmActionsAreDisplayed()).toEqual(4);
     
+  });
+
+  it('should not show secondary navigation when nothing is selected', async () => {
+    const secondaryNavigation: ElementFinder = await page.getSecondaryNavigation();
+    expect<any>(secondaryNavigation.isPresent()).toBe(false);
+  });
+
+  it('should not show secondary navigation items when there is no children on the selected item', async () => {
+
+    // click the first menu item
+    await page.pageHeader2.$$('.horizontal-navigation-button').get(0).click();
+
+    // get the required elements
+    const secondaryNavigation: ElementFinder = await page.getSecondaryNavigation();
+    const tabset = await secondaryNavigation.$('.nav-tabs');
+
+    expect<any>(secondaryNavigation.isPresent()).toBe(true);
+    expect<any>(tabset.isPresent()).toBe(false);
+  });
+
+  it('should show secondary navigation items when there are children on the selected item', async () => {
+
+    // click the second menu item
+    await page.pageHeader2.$$('.horizontal-navigation-button').get(1).click();
+
+    // get the required elements
+    const secondaryNavigation: ElementFinder = await page.getSecondaryNavigation();
+    const tabset = await secondaryNavigation.$('.nav-tabs');
+    const tabs: ElementArrayFinder = await tabset.$$('li');
+
+    expect<any>(secondaryNavigation.isPresent()).toBe(true);
+    expect<any>(tabset.isPresent()).toBe(true);
+
+    // check the number of tabs are correct
+    expect<any>(tabs.length).toBe(3);
+  });
+
+  it('should align the tabset accordingly', async () => {
+
+    // click the second menu item
+    await page.pageHeader2.$$('.horizontal-navigation-button').get(1).click();
+
+    // get the required elements
+    const secondaryNavigation: ElementFinder = await page.getSecondaryNavigation();
+
+    // should contain the 'center' class initially
+    expect(await page.getClasses(secondaryNavigation)).toContain('center');
+
+    // click the left align button
+    await page.alignLeftButton.click();
+
+    // should now contain the 'left' class
+    expect(await page.getClasses(secondaryNavigation)).toContain('left');
+
+    // click the right align button
+    await page.alignRightButton.click();
+
+    // should now contain the 'right' class
+    expect(await page.getClasses(secondaryNavigation)).toContain('right');
+
+    // click the center align button
+    await page.alignCenterButton.click();
+
+    // should now contain the 'center' class
+    expect(await page.getClasses(secondaryNavigation)).toContain('center');
+  });
+
+  it('should call the select functions when secondary items are clicked', async () => {
+
+    // click the second menu item
+    await page.pageHeader2.$$('.horizontal-navigation-button').get(1).click();
+
+    // get the required elements
+    const secondaryNavigation: ElementFinder = await page.getSecondaryNavigation();
+    const tabset = await secondaryNavigation.$('.nav-tabs');
+    const tabs: ElementArrayFinder = await tabset.$$('li');
+
+    // click the first item
+    await tabs[0].click();
+
+    // check the selected text
+    expect(await page.selected.getText()).toBe('Daily View');
+
+    // click the second item
+    await tabs[1].click();
+
+    // check the selected text
+    expect(await page.selected.getText()).toBe('Weekly View');
+
+    // click the third item
+    await tabs[2].click();
+
+    // check the selected text
+    expect(await page.selected.getText()).toBe('Monthly View');
+  });
+
+  it('should automatically select the first item when autoselect is enabled', async () => {
+
+    // enable autoselect
+    await page.autoselectButton.click();
+
+    // click the second menu item
+    await page.pageHeader2.$$('.horizontal-navigation-button').get(1).click();
+
+    // get the required elements
+    const secondaryNavigation: ElementFinder = await page.getSecondaryNavigation();
+    const tabset = await secondaryNavigation.$('.nav-tabs');
+    const tabs: ElementArrayFinder = await tabset.$$('li');
+
+    // check the selected text
+    expect(await page.selected.getText()).toBe('Daily View');
+  });
+
+  it('should not deselect child items when the parent item is clicked again', async () => {
+
+    // enable autoselect
+    await page.autoselectButton.click();
+
+    // click the second menu item
+    await page.pageHeader2.$$('.horizontal-navigation-button').get(1).click();
+
+    // get the required elements
+    const secondaryNavigation: ElementFinder = await page.getSecondaryNavigation();
+    const tabset = await secondaryNavigation.$('.nav-tabs');
+    const tabs: ElementArrayFinder = await tabset.$$('li');
+
+    // click the second item
+    await tabs[1].click();
+
+    // check the selected text
+    expect(await page.selected.getText()).toBe('Weekly View');
+
+    // click the second menu item again
+    await page.pageHeader2.$$('.horizontal-navigation-button').get(1).click();
+
+    // check the selected text has not changed
+    expect(await page.selected.getText()).toBe('Weekly View');
   });
 });
