@@ -1,41 +1,44 @@
-import { Component, Input, ViewChild, ViewChildren, QueryList, OnInit, Output, EventEmitter, ElementRef } from '@angular/core';
+import { Component, ElementRef, Input, OnDestroy, OnInit, QueryList, ViewChild, ViewChildren } from '@angular/core';
 import { BsDropdownDirective } from 'ngx-bootstrap/dropdown';
+import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { Subscription } from 'rxjs/Subscription';
+import { PageHeaderService } from '../../page-header.service';
 import { PageHeaderNavigationDropdownItemComponent } from '../navigation-dropdown-item/navigation-dropdown-item.component';
-import { PageHeaderNavigationItem, PageHeaderNavigationDropdownItem } from '../navigation.component';
+import { PageHeaderNavigationItem } from '../navigation.component';
 
 @Component({
     selector: 'ux-page-header-horizontal-navigation-item',
     templateUrl: './navigation-item.component.html'
 })
-export class PageHeaderNavigationItemComponent implements OnInit {
+export class PageHeaderNavigationItemComponent implements OnInit, OnDestroy {
 
     @ViewChild('menu') menu: BsDropdownDirective;
-    @ViewChildren(PageHeaderNavigationDropdownItemComponent) dropdownComponents: QueryList<PageHeaderNavigationDropdownItemComponent>;
-
+    @ViewChildren(PageHeaderNavigationDropdownItemComponent) dropdowns: QueryList<PageHeaderNavigationDropdownItemComponent>;
+    
     @Input() item: PageHeaderNavigationItem;
-    @Output() onSelect: EventEmitter<PageHeaderNavigationDropdownItem> = new EventEmitter<PageHeaderNavigationDropdownItem>();
+    
+    secondary$: BehaviorSubject<boolean> = this._pageHeaderService.secondary$;
+    
+    private _subscription: Subscription;
 
-    constructor(public elementRef: ElementRef) {}
+    constructor(public elementRef: ElementRef, private _pageHeaderService: PageHeaderService) { }
 
     ngOnInit() {
-        this.menu.onHidden.subscribe(() => this.dropdownComponents.forEach(dropdown => dropdown.close()));
+        this._subscription = this.menu.onHidden.subscribe(() => this.dropdowns.forEach(dropdown => dropdown.close()));
     }
 
-    selectItem() {
+    ngOnDestroy(): void {
+        this._subscription.unsubscribe();
+    }
+
+    select() {
 
         // if the item has children then do nothing at this stage 
-        if (this.item.children) {
+        if (this.item.children && this._pageHeaderService.secondary$.getValue() === false) {
             return;
         }
 
         // otherwise select the current item
-        this.onItemSelect(this.item);
-    }
-
-    onItemSelect(item: PageHeaderNavigationItem | PageHeaderNavigationDropdownItem) {
-        this.onSelect.emit(item);
-
-        // select the current item
-        this.item.selected = true;
+        this._pageHeaderService.select(this.item);
     }
 }
