@@ -1,20 +1,28 @@
-import { Directive, Output, EventEmitter, ElementRef, Input, NgZone, OnInit } from '@angular/core';
-import { ResizeService, ResizeDimensions } from './resize.service';
+import { Directive, ElementRef, EventEmitter, Input, NgZone, OnDestroy, OnInit, Output } from '@angular/core';
+import { Subscription } from 'rxjs/Subscription';
 import { debounceTime } from 'rxjs/operators/debounceTime';
+import { ResizeDimensions, ResizeService } from './resize.service';
 
 @Directive({
-    selector: '[uxResize]'
+    selector: '[uxResize]',
+    providers: [ResizeService]
 })
-export class ResizeDirective implements OnInit {
+export class ResizeDirective implements OnInit, OnDestroy {
 
     @Input() throttle: number = 0;
     @Output() uxResize: EventEmitter<ResizeDimensions> = new EventEmitter<ResizeDimensions>();
 
+    private _subscription: Subscription;
+
     constructor(private _elementRef: ElementRef, private _resizeService: ResizeService, private _ngZone: NgZone) { }
 
     ngOnInit(): void {
-        this._resizeService.addResizeListener(this._elementRef.nativeElement)
+        this._subscription = this._resizeService.addResizeListener(this._elementRef.nativeElement)
             .pipe(debounceTime(this.throttle))
             .subscribe((event: ResizeDimensions) => this._ngZone.run(() => this.uxResize.emit(event)));
+    }
+
+    ngOnDestroy(): void {
+        this._subscription.unsubscribe();
     }
 }
