@@ -520,22 +520,27 @@ describe('treegrid', function () {
     // create a new scope
     $scope = $rootScope.$new();
 
-    $scope.vm = {};
+    // add variables to controller as property
+    $scope.vm = props;
 
-    // iterate each prop and add to scope
+    // prepare the controller
+    var ctrlFn = $controller('TreegridCtrl', {
+      $scope: $scope
+    }, true);
+
+    // add all props to instance (as this component has bindToController: true)
     for (var prop in props) {
-      $scope.vm[prop] = props[prop];
+      ctrlFn.instance[prop] = props[prop];
     }
 
-    var ctrl = $controller('TreegridCtrl', {
-      $scope: $scope
-    }, props);
+    // create the controller
+    var ctrl = ctrlFn();
 
     // perform initial digest
     $scope.$digest();
 
     return ctrl;
-  }
+}
 
   function failTest(error) {
     expect(error).toBeUndefined();
